chore: refactor namespace and resource_definition resources
Refactor meta.Namespace and meta.ResourceDefinition by implementing them using typed.Resource and moving ResourceDefinitionSpec to a side pkg to prevent cyclic imports.
