+++ pkg/resource/meta/namespace.go
@@ -6,15 +6,33 @@ package meta
 
 import (
 	"github.com/cosi-project/runtime/pkg/resource"
+	"github.com/cosi-project/runtime/pkg/resource/typed"
 )
 
 // NamespaceType is the type of Namespace.
 const NamespaceType = resource.Type("Namespaces.meta.cosi.dev")
 
 // Namespace provides metadata about namespaces.
-type Namespace struct {
-	spec NamespaceSpec
-	md   resource.Metadata
+type Namespace = typed.Resource[NamespaceSpec, NamespaceRD]
+
+// NewNamespace initializes a Namespace resource.
+func NewNamespace(id resource.ID, spec NamespaceSpec) *Namespace {
+	return typed.NewResource[NamespaceSpec, NamespaceRD](
+		resource.NewMetadata(NamespaceName, NamespaceType, id, resource.VersionUndefined),
+		spec,
+	)
+}
+
+// NamespaceRD provides auxiliary methods for Namespace.
+type NamespaceRD struct{}
+
+// ResourceDefinition implements core.ResourceDefinitionProvider interface.
+func (NamespaceRD) ResourceDefinition(_ resource.Metadata, _ NamespaceSpec) ResourceDefinitionSpec {
+	return ResourceDefinitionSpec{
+		Type:             NamespaceType,
+		DefaultNamespace: NamespaceName,
+		Aliases:          []resource.Type{"ns"},
+	}
 }
 
 // NamespaceSpec provides Namespace definition.
@@ -22,41 +40,7 @@ type NamespaceSpec struct {
 	Description string `yaml:"description"`
 }
 
-// NewNamespace initializes a Namespace resource.
-func NewNamespace(id resource.ID, spec NamespaceSpec) *Namespace {
-	r := &Namespace{
-		md:   resource.NewMetadata(NamespaceName, NamespaceType, id, resource.VersionUndefined),
-		spec: spec,
-	}
-
-	r.md.BumpVersion()
-
-	return r
-}
-
-// Metadata implements resource.Resource.
-func (r *Namespace) Metadata() *resource.Metadata {
-	return &r.md
-}
-
-// Spec implements resource.Resource.
-func (r *Namespace) Spec() interface{} {
-	return r.spec
-}
-
-// DeepCopy implements resource.Resource.
-func (r *Namespace) DeepCopy() resource.Resource { //nolint:ireturn
-	return &Namespace{
-		md:   r.md,
-		spec: r.spec,
-	}
-}
-
-// ResourceDefinition implements core.ResourceDefinitionProvider interface.
-func (r *Namespace) ResourceDefinition() ResourceDefinitionSpec {
-	return ResourceDefinitionSpec{
-		Type:             NamespaceType,
-		DefaultNamespace: NamespaceName,
-		Aliases:          []resource.Type{"ns"},
-	}
+// DeepCopy generates a deep copy of NamespaceSpec.
+func (n NamespaceSpec) DeepCopy() NamespaceSpec {
+	return n
 }
